use u32 in GraphMutation

+++ iris-mpc-cpu/src/hnsw/graph/mutation.rs
@@ -10,14 +10,14 @@ pub enum GraphMutation<Vector: Ord> {
     // Add bidirectional links.
     InsertNode {
         // List of layer, neighbors.
-        layers: Vec<(usize, Vec<Vector>)>,
+        layers: Vec<(u32, Vec<Vector>)>,
         update_ep: UpdateEntryPoint,
         id: Vector,
     },
     // Compact a neighborhood.
     Compact {
         to_remove: Vec<Vector>,
-        layer: usize,
+        layer: u32,
         id: Vector,
     },
 }
